api/requirements.txt
fastapi
uvicorn[standard]
Pillow
pydantic
numpy
requests
